Cache Maven dependencies in the GitHub Actions workflow
Also bumps to actions/cache@v2, and only cleanup before cache
if the cache wasn't hit.

For the Maven dependencies, make sure SNAPSHOT artifacts
are never cached, as well as the LOCAL artifacts, used
to give Maven the results of the Gradle build for the
J2Cl tests.
The cache key includes the Maven version, as its super-pom
selects the default versions for many plugins.